[REFACTOR] 아토믹 디자인 리팩토링 TEMPLATE 추가
