+++ src/components/atoms/Group/flex.tsx
@@ -1,10 +1,10 @@
 import React from "react";
-import * as Styled from "./index.styles";
-import { Props } from "./index.types";
+import styled from "@emotion/styled";
+import { FlexProps, Props } from "./flex.types";
 
 const Flex = React.forwardRef(function Flex(
   {
-    as = "div",
+    as = "span",
     display = "flex",
     direction = "row",
     wrap = "nowrap",
@@ -12,14 +12,13 @@ const Flex = React.forwardRef(function Flex(
     alignItems = "flex-start",
     alignContent = "normal",
     backgroundColor = "inherit",
-    gap = 0,
     children,
     ...rest
   }: React.PropsWithChildren<Props>,
   forwardedRef: React.Ref<HTMLElement>
 ) {
   return (
-    <Styled.Flex
+    <Wrapper
       as={as}
       display={display}
       direction={direction}
@@ -28,13 +27,23 @@ const Flex = React.forwardRef(function Flex(
       alignItems={alignItems}
       alignContent={alignContent}
       backgroundColor={backgroundColor}
-      gap={gap}
       ref={forwardedRef}
       {...rest}
     >
       {children}
-    </Styled.Flex>
+    </Wrapper>
   );
 });
 
 export default Flex;
+
+export const Wrapper = styled.span<Required<FlexProps>>`
+  display: ${({ display }) => display};
+  flex-direction: ${({ direction }) => direction};
+  flex-wrap: ${({ wrap }) => wrap};
+  justify-content: ${({ justifyContent }) => justifyContent};
+  align-items: ${({ alignItems }) => alignItems};
+  align-content: ${({ alignContent }) => alignContent};
+  background-color: ${({ backgroundColor, theme }) =>
+    theme.colors[backgroundColor]};
+`;
