+++ src/components/templates/mono-template.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const MonoTemplate = () => {
+  return <div>mono-template</div>;
+};
+
+export default MonoTemplate;
