+++ src/pages/Auth/SignUpForm.tsx
@@ -11,10 +11,10 @@ import { RiLockPasswordLine } from "react-icons/ri";
 import { toast } from "react-toastify";
 
 import TextFiled from "@components/molecules/TextField";
-import Button from "Button";
-import Flex from "@components/ui/Flex";
 import { emailRule, krRule, passwordRule } from "@utils/validate";
 import userApis from "@api/modules/user.api";
+import RectangleButton from "@components/atoms/Button/rectangle-button";
+import Flex from "@components/atoms/Group/flex";
 
 type FormValues = {
   displayName: string;
@@ -159,9 +159,9 @@ const SignUpForm = () => {
           helperIcon={<AiOutlineExclamationCircle />}
           {...register("confirmPassword", formState.confirmPassword)}
         />
-        <Button type="submit" variant="green">
+        <RectangleButton type="submit" variant="green">
           확인
-        </Button>
+        </RectangleButton>
       </Flex>
     </form>
   );
