+++ src/pages/Follow/FollowUserList.tsx
@@ -1,11 +1,13 @@
 import React from "react";
 import { SwiperSlide } from "swiper/react";
 
-import CustomSiper from "@components/CustomSwiper";
+import { routesGen } from "@routes/router";
+
+import CustomSiper from "@components/molecules/Carousel/CustomSwiper";
 import Box from "@components/atoms/Box";
 import Typography from "@components/atoms/Typograph";
 import Image from "@components/atoms/Image";
-import { routesGen } from "@routes/router";
+
 import followList from "../../db/follow.json";
 
 const FollowUserList = () => {
