+++ src/pages/Home/Page.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { css } from "@emotion/react";
 
-import Button from "@components/atoms/Button/Button";
-import Flex from "@components/atoms/Flex";
+import Button from "@components/atoms/Button/rectangle-button";
+import Flex from "@components/atoms/Group/flex";
 /**
  * Page의 관심사 분리 예시 입니다.
  *
@@ -13,7 +13,7 @@ import Flex from "@components/atoms/Flex";
 
 const Home = () => {
   return (
-    <Flex>
+    <Flex direction="column">
       <Button size="sm" variant="green">
         재사용 버튼
       </Button>
