+++ src/pages/MyPage/EditPageRequire.tsx
@@ -8,9 +8,9 @@ import {
 import { RiLockPasswordLine } from "react-icons/ri";
 import { toast } from "react-toastify";
 import TextFiled from "@components/molecules/TextField";
-import Button from "Button";
-import Flex from "@components/ui/Flex";
 import { emailRule, krEgNumRule, passwordRule } from "@utils/validate";
+import Flex from "@components/atoms/Group/flex";
+import Button from "@components/atoms/Button/rectangle-button";
 
 type FormValues = {
   nickName: string;
